Refactor FFD
Default FFD application source is now the minimal reference design,
consisting of mics input -> avona dsp pipeline -> wanson inference
engine -> xscope print.

All alternative outputs/modes/peripheral IO has been moved to the
extensions, in the ext folder.

The purpose for this change is for the minimal reference application to
not be burdened with various XMOS specific configurations and #defines.