Updated appconfINFERENCE_RESET_DELAY_MS
Increased the value to give a bit more room if a user
is inclined to wait for feedback on the wake word detection.

In the case of audio playback, the user may also wait for playback
to finish during the STATE_PROCESSING_COMMAND phase. A
slightly longer time is given in such configurations.