Fix spurious Linux usb audio errors with ffd
It was found that on Ubuntu there were inconsistent usb related issues
in the debug build of ffd.

It was found that there were timing related issues, along with a few
other minor ones.  The solution to resolve the potential deadlock that
can occur from these race conditions was to put the usb_audio intertile
traffic on a dedicated intertile driver.