Fixed low_power_audio_buffer_enqueue()
Fixed guard clause. The ring buffer should not process a dequeue
request if there are less than appconfAUDIO_PIPELINE_FRAME_ADVANCE
frames in the buffer.

Added an assert for preventing use-cases where an application wants
to enqueue more data than the internal buffer size in one call.

The "size" field of the buffer is in bytes. This field was being
handled inconsistently within the function.