Added CI workflow for releases
Adjustments to existing workflows to be callable. These workflows
now accept an input argument to specify the artifact names that will
be uploaded during workflow execution.

If the input is empty, then the workflow defaults to artifact names
specified in the yaml file. This allows the workflow to be
individually dispatched with similar behavior as before.

Releases are triggered by ci.yml when pushing tags with the
appropriate format. A release run executes a different sequence of
jobs and will produce artifacts names with versions appended.